+++ IPED_Gui/MainForm.Designer.cs
@@ -31,7 +31,6 @@
             components = new System.ComponentModel.Container();
             GroupBox groupBox2;
             System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(MainForm));
-            GroupBox groupBox1;
             Button buttonAudioTranslationInputDirectory;
             Button buttonAudioTranslationOutputDirectory;
             Label label3;
@@ -42,12 +41,40 @@
             Button buttonSettingsIndexTemp;
             Button buttonSettingsHashesDB;
             Button buttonSettingsPluginFolder;
+            GroupBox groupBox3;
+            Button buttonAudioTranslationArgosTranslateDirectory;
+            Button buttonAudioTranslationFasterWhisperDirectory;
+            Button buttonAudioTranslationProcessingDirectory;
+            Label label11;
+            Button buttonAudioTranslationServiceProgram;
+            GroupBox groupBox4;
+            Button buttonImageClassificationMobileNetDirectory;
+            Button buttonImageClassificationProcessingDirectory;
+            Label label20;
+            Button buttonImageClassificationServiceProgram;
             checkBox_ProjektErweitern = new CheckBox();
             checkBox_Fortsetzen = new CheckBox();
             checkBox_Portabel = new CheckBox();
             checkBox_InternetdatenLaden = new CheckBox();
-            checkBoxIPEDConfigEnableAudioTranslation = new CheckBox();
-            checkBoxIPEDConfigEnableImageClassification = new CheckBox();
+            checkBoxAudioTranslationUseGPU = new CheckBox();
+            buttonAudioTranslationStartProcess = new Button();
+            comboBoxAudioTranslationModelSize = new ComboBox();
+            label15 = new Label();
+            label14 = new Label();
+            textBoxAudioTranslationArgosTranslateDirectory = new TextBox();
+            label13 = new Label();
+            textBoxAudioTranslationFasterWhisperDirectory = new TextBox();
+            label12 = new Label();
+            textBoxAudioTranslationProcessingDirectory = new TextBox();
+            textBoxAudioTranslationServiceProgram = new TextBox();
+            buttonImageClassificationStartProcess = new Button();
+            comboBoxImageClassificationLanguage = new ComboBox();
+            label16 = new Label();
+            label18 = new Label();
+            textBoxImageClassificationMobileNetDirectory = new TextBox();
+            label19 = new Label();
+            textBoxImageClassificationProcessingDirectory = new TextBox();
+            textBoxImageClassificationServiceProgram = new TextBox();
             label_Spuren = new Label();
             listBox_Spuren = new ListBox();
             button_VerzeichnisHinzufuegen = new Button();
@@ -64,6 +91,8 @@
             label1 = new Label();
             label2 = new Label();
             checkBoxAudioTranslationProcessVideos = new CheckBox();
+            checkBoxIPEDConfigEnableAudioTranslation = new CheckBox();
+            checkBoxIPEDConfigEnableImageClassification = new CheckBox();
             tabControl1 = new TabControl();
             tabPageOverview = new TabPage();
             tabPageSettings = new TabPage();
@@ -91,8 +120,9 @@
             openFileDialogSettingsIpedExePath = new OpenFileDialog();
             folderBrowserDialog = new FolderBrowserDialog();
             openFileDialogSettingsHashesDB = new OpenFileDialog();
+            openFileDialogAudioTranslationServiceProgram = new OpenFileDialog();
+            openFileDialogImageClassificationServiceProgram = new OpenFileDialog();
             groupBox2 = new GroupBox();
-            groupBox1 = new GroupBox();
             buttonAudioTranslationInputDirectory = new Button();
             buttonAudioTranslationOutputDirectory = new Button();
             label3 = new Label();
@@ -103,8 +133,20 @@
             buttonSettingsIndexTemp = new Button();
             buttonSettingsHashesDB = new Button();
             buttonSettingsPluginFolder = new Button();
+            groupBox3 = new GroupBox();
+            buttonAudioTranslationArgosTranslateDirectory = new Button();
+            buttonAudioTranslationFasterWhisperDirectory = new Button();
+            buttonAudioTranslationProcessingDirectory = new Button();
+            label11 = new Label();
+            buttonAudioTranslationServiceProgram = new Button();
+            groupBox4 = new GroupBox();
+            buttonImageClassificationMobileNetDirectory = new Button();
+            buttonImageClassificationProcessingDirectory = new Button();
+            label20 = new Label();
+            buttonImageClassificationServiceProgram = new Button();
             groupBox2.SuspendLayout();
-            groupBox1.SuspendLayout();
+            groupBox3.SuspendLayout();
+            groupBox4.SuspendLayout();
             tableLayoutPanel1.SuspendLayout();
             tabControl1.SuspendLayout();
             tabPageOverview.SuspendLayout();
@@ -116,6 +158,7 @@
             // 
             // groupBox2
             // 
+            tableLayoutPanel1.SetColumnSpan(groupBox2, 2);
             groupBox2.Controls.Add(checkBox_ProjektErweitern);
             groupBox2.Controls.Add(checkBox_Fortsetzen);
             groupBox2.Controls.Add(checkBox_Portabel);
@@ -123,7 +166,7 @@
             groupBox2.Dock = DockStyle.Fill;
             groupBox2.Location = new Point(163, 253);
             groupBox2.Name = "groupBox2";
-            groupBox2.Size = new Size(326, 246);
+            groupBox2.Size = new Size(658, 246);
             groupBox2.TabIndex = 20;
             groupBox2.TabStop = false;
             groupBox2.Text = "Optionen";
@@ -143,7 +186,7 @@
             // checkBox_Fortsetzen
             // 
             checkBox_Fortsetzen.AutoSize = true;
-            checkBox_Fortsetzen.Location = new Point(6, 47);
+            checkBox_Fortsetzen.Location = new Point(6, 72);
             checkBox_Fortsetzen.Name = "checkBox_Fortsetzen";
             checkBox_Fortsetzen.Size = new Size(158, 19);
             checkBox_Fortsetzen.TabIndex = 9;
@@ -155,7 +198,7 @@
             // checkBox_Portabel
             // 
             checkBox_Portabel.AutoSize = true;
-            checkBox_Portabel.Location = new Point(6, 72);
+            checkBox_Portabel.Location = new Point(6, 122);
             checkBox_Portabel.Name = "checkBox_Portabel";
             checkBox_Portabel.Size = new Size(70, 19);
             checkBox_Portabel.TabIndex = 10;
@@ -167,7 +210,7 @@
             // checkBox_InternetdatenLaden
             // 
             checkBox_InternetdatenLaden.AutoSize = true;
-            checkBox_InternetdatenLaden.Location = new Point(6, 97);
+            checkBox_InternetdatenLaden.Location = new Point(6, 147);
             checkBox_InternetdatenLaden.Name = "checkBox_InternetdatenLaden";
             checkBox_InternetdatenLaden.Size = new Size(129, 19);
             checkBox_InternetdatenLaden.TabIndex = 11;
@@ -176,46 +219,10 @@
             checkBox_InternetdatenLaden.UseVisualStyleBackColor = true;
             checkBox_InternetdatenLaden.CheckedChanged += checkBox_InternetdatenLaden_CheckedChanged;
             // 
-            // groupBox1
-            // 
-            groupBox1.Controls.Add(checkBoxIPEDConfigEnableAudioTranslation);
-            groupBox1.Controls.Add(checkBoxIPEDConfigEnableImageClassification);
-            groupBox1.Dock = DockStyle.Fill;
-            groupBox1.Location = new Point(495, 253);
-            groupBox1.Name = "groupBox1";
-            groupBox1.Size = new Size(326, 246);
-            groupBox1.TabIndex = 21;
-            groupBox1.TabStop = false;
-            groupBox1.Text = "Funktionen";
-            // 
-            // checkBoxIPEDConfigEnableAudioTranslation
-            // 
-            checkBoxIPEDConfigEnableAudioTranslation.AutoSize = true;
-            checkBoxIPEDConfigEnableAudioTranslation.Location = new Point(6, 22);
-            checkBoxIPEDConfigEnableAudioTranslation.Name = "checkBoxIPEDConfigEnableAudioTranslation";
-            checkBoxIPEDConfigEnableAudioTranslation.Size = new Size(222, 19);
-            checkBoxIPEDConfigEnableAudioTranslation.TabIndex = 12;
-            checkBoxIPEDConfigEnableAudioTranslation.Text = "Audio Transkription und Übersetzung";
-            toolTip1.SetToolTip(checkBoxIPEDConfigEnableAudioTranslation, "Führt bei Audio- und Videodateien Transkriptionen und Übersetzung ins Deutsche mit Hilfe des Auswerteservers durch.");
-            checkBoxIPEDConfigEnableAudioTranslation.UseVisualStyleBackColor = true;
-            checkBoxIPEDConfigEnableAudioTranslation.CheckedChanged += checkBoxIPEDConfigEnabelAudioTranslation_CheckedChanged;
-            // 
-            // checkBoxIPEDConfigEnableImageClassification
-            // 
-            checkBoxIPEDConfigEnableImageClassification.AutoSize = true;
-            checkBoxIPEDConfigEnableImageClassification.Location = new Point(6, 47);
-            checkBoxIPEDConfigEnableImageClassification.Name = "checkBoxIPEDConfigEnableImageClassification";
-            checkBoxIPEDConfigEnableImageClassification.Size = new Size(125, 19);
-            checkBoxIPEDConfigEnableImageClassification.TabIndex = 13;
-            checkBoxIPEDConfigEnableImageClassification.Text = "Bilder klassifizieren";
-            toolTip1.SetToolTip(checkBoxIPEDConfigEnableImageClassification, "Führt eine Klassifizierung von Bilddateien mit Hilfe des Auswerteservers durch.");
-            checkBoxIPEDConfigEnableImageClassification.UseVisualStyleBackColor = true;
-            checkBoxIPEDConfigEnableImageClassification.CheckedChanged += checkBoxIPEDConfigEnableImageClassification_CheckedChanged;
-            // 
             // buttonAudioTranslationInputDirectory
             // 
             buttonAudioTranslationInputDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
-            buttonAudioTranslationInputDirectory.Location = new Point(827, 39);
+            buttonAudioTranslationInputDirectory.Location = new Point(826, 67);
             buttonAudioTranslationInputDirectory.Name = "buttonAudioTranslationInputDirectory";
             buttonAudioTranslationInputDirectory.Size = new Size(120, 23);
             buttonAudioTranslationInputDirectory.TabIndex = 9;
@@ -226,7 +233,7 @@
             // buttonAudioTranslationOutputDirectory
             // 
             buttonAudioTranslationOutputDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
-            buttonAudioTranslationOutputDirectory.Location = new Point(827, 69);
+            buttonAudioTranslationOutputDirectory.Location = new Point(826, 97);
             buttonAudioTranslationOutputDirectory.Name = "buttonAudioTranslationOutputDirectory";
             buttonAudioTranslationOutputDirectory.Size = new Size(120, 23);
             buttonAudioTranslationOutputDirectory.TabIndex = 12;
@@ -237,7 +244,7 @@
             // label3
             // 
             label3.AutoSize = true;
-            label3.Location = new Point(43, 72);
+            label3.Location = new Point(45, 100);
             label3.Name = "label3";
             label3.Size = new Size(111, 15);
             label3.TabIndex = 16;
@@ -247,7 +254,7 @@
             // buttonImageClassificationOutputDirectory
             // 
             buttonImageClassificationOutputDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
-            buttonImageClassificationOutputDirectory.Location = new Point(827, 69);
+            buttonImageClassificationOutputDirectory.Location = new Point(829, 97);
             buttonImageClassificationOutputDirectory.Name = "buttonImageClassificationOutputDirectory";
             buttonImageClassificationOutputDirectory.Size = new Size(120, 23);
             buttonImageClassificationOutputDirectory.TabIndex = 18;
@@ -258,7 +265,7 @@
             // buttonImageClassificationInputDirectory
             // 
             buttonImageClassificationInputDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
-            buttonImageClassificationInputDirectory.Location = new Point(827, 40);
+            buttonImageClassificationInputDirectory.Location = new Point(829, 68);
             buttonImageClassificationInputDirectory.Name = "buttonImageClassificationInputDirectory";
             buttonImageClassificationInputDirectory.Size = new Size(120, 23);
             buttonImageClassificationInputDirectory.TabIndex = 15;
@@ -269,7 +276,7 @@
             // label4
             // 
             label4.AutoSize = true;
-            label4.Location = new Point(43, 43);
+            label4.Location = new Point(45, 71);
             label4.Name = "label4";
             label4.Size = new Size(107, 15);
             label4.TabIndex = 19;
@@ -319,6 +326,329 @@
             buttonSettingsPluginFolder.Text = "Auswählen ...";
             buttonSettingsPluginFolder.UseVisualStyleBackColor = true;
             buttonSettingsPluginFolder.Click += buttonSettingsPluginFolder_Click;
+            // 
+            // groupBox3
+            // 
+            groupBox3.Controls.Add(checkBoxAudioTranslationUseGPU);
+            groupBox3.Controls.Add(buttonAudioTranslationStartProcess);
+            groupBox3.Controls.Add(comboBoxAudioTranslationModelSize);
+            groupBox3.Controls.Add(label15);
+            groupBox3.Controls.Add(label14);
+            groupBox3.Controls.Add(textBoxAudioTranslationArgosTranslateDirectory);
+            groupBox3.Controls.Add(buttonAudioTranslationArgosTranslateDirectory);
+            groupBox3.Controls.Add(label13);
+            groupBox3.Controls.Add(textBoxAudioTranslationFasterWhisperDirectory);
+            groupBox3.Controls.Add(buttonAudioTranslationFasterWhisperDirectory);
+            groupBox3.Controls.Add(label12);
+            groupBox3.Controls.Add(textBoxAudioTranslationProcessingDirectory);
+            groupBox3.Controls.Add(buttonAudioTranslationProcessingDirectory);
+            groupBox3.Controls.Add(label11);
+            groupBox3.Controls.Add(textBoxAudioTranslationServiceProgram);
+            groupBox3.Controls.Add(buttonAudioTranslationServiceProgram);
+            groupBox3.Location = new Point(3, 151);
+            groupBox3.Name = "groupBox3";
+            groupBox3.Size = new Size(984, 386);
+            groupBox3.TabIndex = 15;
+            groupBox3.TabStop = false;
+            groupBox3.Text = "Lokale Verarbeitung";
+            // 
+            // checkBoxAudioTranslationUseGPU
+            // 
+            checkBoxAudioTranslationUseGPU.AutoSize = true;
+            checkBoxAudioTranslationUseGPU.Location = new Point(196, 167);
+            checkBoxAudioTranslationUseGPU.Name = "checkBoxAudioTranslationUseGPU";
+            checkBoxAudioTranslationUseGPU.Size = new Size(110, 19);
+            checkBoxAudioTranslationUseGPU.TabIndex = 16;
+            checkBoxAudioTranslationUseGPU.Text = "GPU verwenden";
+            toolTip1.SetToolTip(checkBoxAudioTranslationUseGPU, "Verwenden Sie diese Option, wenn Ihr PC über eine Grafikkarte mit mindestens 8GB Video-RAM verfügt");
+            checkBoxAudioTranslationUseGPU.UseVisualStyleBackColor = true;
+            checkBoxAudioTranslationUseGPU.CheckedChanged += checkBoxAudioTranslationUseGPU_CheckedChanged;
+            // 
+            // buttonAudioTranslationStartProcess
+            // 
+            buttonAudioTranslationStartProcess.Anchor = AnchorStyles.Bottom | AnchorStyles.Left | AnchorStyles.Right;
+            buttonAudioTranslationStartProcess.Location = new Point(196, 192);
+            buttonAudioTranslationStartProcess.Name = "buttonAudioTranslationStartProcess";
+            buttonAudioTranslationStartProcess.Size = new Size(154, 23);
+            buttonAudioTranslationStartProcess.TabIndex = 30;
+            buttonAudioTranslationStartProcess.Text = "Prozess starten";
+            buttonAudioTranslationStartProcess.UseVisualStyleBackColor = true;
+            buttonAudioTranslationStartProcess.Click += buttonAudioTranslationStartProcess_Click;
+            // 
+            // comboBoxAudioTranslationModelSize
+            // 
+            comboBoxAudioTranslationModelSize.FormattingEnabled = true;
+            comboBoxAudioTranslationModelSize.Items.AddRange(new object[] { "tiny", "base", "small", "medium", "large-v2" });
+            comboBoxAudioTranslationModelSize.Location = new Point(196, 138);
+            comboBoxAudioTranslationModelSize.Name = "comboBoxAudioTranslationModelSize";
+            comboBoxAudioTranslationModelSize.Size = new Size(121, 23);
+            comboBoxAudioTranslationModelSize.TabIndex = 29;
+            comboBoxAudioTranslationModelSize.SelectedIndexChanged += comboBoxAudioTranslationModelSize_SelectedIndexChanged;
+            // 
+            // label15
+            // 
+            label15.AutoSize = true;
+            label15.Location = new Point(39, 141);
+            label15.Name = "label15";
+            label15.Size = new Size(75, 15);
+            label15.TabIndex = 28;
+            label15.Text = "Modellgröße";
+            toolTip1.SetToolTip(label15, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // label14
+            // 
+            label14.AutoSize = true;
+            label14.Location = new Point(39, 112);
+            label14.Name = "label14";
+            label14.Size = new Size(148, 15);
+            label14.TabIndex = 25;
+            label14.Text = "Argos Translate Verzeichnis";
+            toolTip1.SetToolTip(label14, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // textBoxAudioTranslationArgosTranslateDirectory
+            // 
+            textBoxAudioTranslationArgosTranslateDirectory.Location = new Point(196, 109);
+            textBoxAudioTranslationArgosTranslateDirectory.Name = "textBoxAudioTranslationArgosTranslateDirectory";
+            textBoxAudioTranslationArgosTranslateDirectory.ReadOnly = true;
+            textBoxAudioTranslationArgosTranslateDirectory.Size = new Size(621, 23);
+            textBoxAudioTranslationArgosTranslateDirectory.TabIndex = 26;
+            // 
+            // buttonAudioTranslationArgosTranslateDirectory
+            // 
+            buttonAudioTranslationArgosTranslateDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
+            buttonAudioTranslationArgosTranslateDirectory.Location = new Point(823, 108);
+            buttonAudioTranslationArgosTranslateDirectory.Name = "buttonAudioTranslationArgosTranslateDirectory";
+            buttonAudioTranslationArgosTranslateDirectory.Size = new Size(120, 23);
+            buttonAudioTranslationArgosTranslateDirectory.TabIndex = 27;
+            buttonAudioTranslationArgosTranslateDirectory.Text = "Auswählen ...";
+            buttonAudioTranslationArgosTranslateDirectory.UseVisualStyleBackColor = true;
+            buttonAudioTranslationArgosTranslateDirectory.Click += buttonAudioTranslationArgosTranslateDirectory_Click;
+            // 
+            // label13
+            // 
+            label13.AutoSize = true;
+            label13.Location = new Point(39, 83);
+            label13.Name = "label13";
+            label13.Size = new Size(145, 15);
+            label13.TabIndex = 22;
+            label13.Text = "Faster Whisper Verzeichnis";
+            toolTip1.SetToolTip(label13, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // textBoxAudioTranslationFasterWhisperDirectory
+            // 
+            textBoxAudioTranslationFasterWhisperDirectory.Location = new Point(196, 80);
+            textBoxAudioTranslationFasterWhisperDirectory.Name = "textBoxAudioTranslationFasterWhisperDirectory";
+            textBoxAudioTranslationFasterWhisperDirectory.ReadOnly = true;
+            textBoxAudioTranslationFasterWhisperDirectory.Size = new Size(621, 23);
+            textBoxAudioTranslationFasterWhisperDirectory.TabIndex = 23;
+            // 
+            // buttonAudioTranslationFasterWhisperDirectory
+            // 
+            buttonAudioTranslationFasterWhisperDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
+            buttonAudioTranslationFasterWhisperDirectory.Location = new Point(823, 79);
+            buttonAudioTranslationFasterWhisperDirectory.Name = "buttonAudioTranslationFasterWhisperDirectory";
+            buttonAudioTranslationFasterWhisperDirectory.Size = new Size(120, 23);
+            buttonAudioTranslationFasterWhisperDirectory.TabIndex = 24;
+            buttonAudioTranslationFasterWhisperDirectory.Text = "Auswählen ...";
+            buttonAudioTranslationFasterWhisperDirectory.UseVisualStyleBackColor = true;
+            buttonAudioTranslationFasterWhisperDirectory.Click += buttonAudioTranslationFasterWhisperDirectory_Click;
+            // 
+            // label12
+            // 
+            label12.AutoSize = true;
+            label12.Location = new Point(39, 54);
+            label12.Name = "label12";
+            label12.Size = new Size(137, 15);
+            label12.TabIndex = 19;
+            label12.Text = "Verarbeitungsverzeichnis";
+            toolTip1.SetToolTip(label12, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // textBoxAudioTranslationProcessingDirectory
+            // 
+            textBoxAudioTranslationProcessingDirectory.Location = new Point(196, 51);
+            textBoxAudioTranslationProcessingDirectory.Name = "textBoxAudioTranslationProcessingDirectory";
+            textBoxAudioTranslationProcessingDirectory.ReadOnly = true;
+            textBoxAudioTranslationProcessingDirectory.Size = new Size(621, 23);
+            textBoxAudioTranslationProcessingDirectory.TabIndex = 20;
+            // 
+            // buttonAudioTranslationProcessingDirectory
+            // 
+            buttonAudioTranslationProcessingDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
+            buttonAudioTranslationProcessingDirectory.Location = new Point(823, 50);
+            buttonAudioTranslationProcessingDirectory.Name = "buttonAudioTranslationProcessingDirectory";
+            buttonAudioTranslationProcessingDirectory.Size = new Size(120, 23);
+            buttonAudioTranslationProcessingDirectory.TabIndex = 21;
+            buttonAudioTranslationProcessingDirectory.Text = "Auswählen ...";
+            buttonAudioTranslationProcessingDirectory.UseVisualStyleBackColor = true;
+            buttonAudioTranslationProcessingDirectory.Click += buttonAudioTranslationProcessingDirectory_Click;
+            // 
+            // label11
+            // 
+            label11.AutoSize = true;
+            label11.Location = new Point(39, 25);
+            label11.Name = "label11";
+            label11.Size = new Size(97, 15);
+            label11.TabIndex = 16;
+            label11.Text = "Dienstprogramm";
+            toolTip1.SetToolTip(label11, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // textBoxAudioTranslationServiceProgram
+            // 
+            textBoxAudioTranslationServiceProgram.Location = new Point(196, 22);
+            textBoxAudioTranslationServiceProgram.Name = "textBoxAudioTranslationServiceProgram";
+            textBoxAudioTranslationServiceProgram.ReadOnly = true;
+            textBoxAudioTranslationServiceProgram.Size = new Size(621, 23);
+            textBoxAudioTranslationServiceProgram.TabIndex = 17;
+            // 
+            // buttonAudioTranslationServiceProgram
+            // 
+            buttonAudioTranslationServiceProgram.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
+            buttonAudioTranslationServiceProgram.Location = new Point(823, 21);
+            buttonAudioTranslationServiceProgram.Name = "buttonAudioTranslationServiceProgram";
+            buttonAudioTranslationServiceProgram.Size = new Size(120, 23);
+            buttonAudioTranslationServiceProgram.TabIndex = 18;
+            buttonAudioTranslationServiceProgram.Text = "Auswählen ...";
+            buttonAudioTranslationServiceProgram.UseVisualStyleBackColor = true;
+            buttonAudioTranslationServiceProgram.Click += buttonAudioTranslationServiceProgram_Click;
+            // 
+            // groupBox4
+            // 
+            groupBox4.Controls.Add(buttonImageClassificationStartProcess);
+            groupBox4.Controls.Add(comboBoxImageClassificationLanguage);
+            groupBox4.Controls.Add(label16);
+            groupBox4.Controls.Add(label18);
+            groupBox4.Controls.Add(textBoxImageClassificationMobileNetDirectory);
+            groupBox4.Controls.Add(buttonImageClassificationMobileNetDirectory);
+            groupBox4.Controls.Add(label19);
+            groupBox4.Controls.Add(textBoxImageClassificationProcessingDirectory);
+            groupBox4.Controls.Add(buttonImageClassificationProcessingDirectory);
+            groupBox4.Controls.Add(label20);
+            groupBox4.Controls.Add(textBoxImageClassificationServiceProgram);
+            groupBox4.Controls.Add(buttonImageClassificationServiceProgram);
+            groupBox4.Location = new Point(3, 151);
+            groupBox4.Name = "groupBox4";
+            groupBox4.Size = new Size(984, 386);
+            groupBox4.TabIndex = 21;
+            groupBox4.TabStop = false;
+            groupBox4.Text = "Lokale Verarbeitung";
+            // 
+            // buttonImageClassificationStartProcess
+            // 
+            buttonImageClassificationStartProcess.Anchor = AnchorStyles.Bottom | AnchorStyles.Left | AnchorStyles.Right;
+            buttonImageClassificationStartProcess.Location = new Point(196, 138);
+            buttonImageClassificationStartProcess.Name = "buttonImageClassificationStartProcess";
+            buttonImageClassificationStartProcess.Size = new Size(154, 23);
+            buttonImageClassificationStartProcess.TabIndex = 31;
+            buttonImageClassificationStartProcess.Text = "Prozess starten";
+            buttonImageClassificationStartProcess.UseVisualStyleBackColor = true;
+            buttonImageClassificationStartProcess.Click += buttonImageClassificationStartProcess_Click;
+            // 
+            // comboBoxImageClassificationLanguage
+            // 
+            comboBoxImageClassificationLanguage.FormattingEnabled = true;
+            comboBoxImageClassificationLanguage.Items.AddRange(new object[] { "de", "en" });
+            comboBoxImageClassificationLanguage.Location = new Point(196, 109);
+            comboBoxImageClassificationLanguage.Name = "comboBoxImageClassificationLanguage";
+            comboBoxImageClassificationLanguage.Size = new Size(121, 23);
+            comboBoxImageClassificationLanguage.TabIndex = 29;
+            comboBoxImageClassificationLanguage.SelectedIndexChanged += comboBoxImageClassificationLanguage_SelectedIndexChanged;
+            // 
+            // label16
+            // 
+            label16.AutoSize = true;
+            label16.Location = new Point(39, 112);
+            label16.Name = "label16";
+            label16.Size = new Size(49, 15);
+            label16.TabIndex = 28;
+            label16.Text = "Sprache";
+            toolTip1.SetToolTip(label16, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // label18
+            // 
+            label18.AutoSize = true;
+            label18.Location = new Point(39, 83);
+            label18.Name = "label18";
+            label18.Size = new Size(124, 15);
+            label18.TabIndex = 22;
+            label18.Text = "MobileNet Verzeichnis";
+            toolTip1.SetToolTip(label18, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // textBoxImageClassificationMobileNetDirectory
+            // 
+            textBoxImageClassificationMobileNetDirectory.Location = new Point(196, 80);
+            textBoxImageClassificationMobileNetDirectory.Name = "textBoxImageClassificationMobileNetDirectory";
+            textBoxImageClassificationMobileNetDirectory.ReadOnly = true;
+            textBoxImageClassificationMobileNetDirectory.Size = new Size(621, 23);
+            textBoxImageClassificationMobileNetDirectory.TabIndex = 23;
+            // 
+            // buttonImageClassificationMobileNetDirectory
+            // 
+            buttonImageClassificationMobileNetDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
+            buttonImageClassificationMobileNetDirectory.Location = new Point(823, 79);
+            buttonImageClassificationMobileNetDirectory.Name = "buttonImageClassificationMobileNetDirectory";
+            buttonImageClassificationMobileNetDirectory.Size = new Size(123, 23);
+            buttonImageClassificationMobileNetDirectory.TabIndex = 24;
+            buttonImageClassificationMobileNetDirectory.Text = "Auswählen ...";
+            buttonImageClassificationMobileNetDirectory.UseVisualStyleBackColor = true;
+            buttonImageClassificationMobileNetDirectory.Click += buttonImageClassificationMobileNetDirectory_Click;
+            // 
+            // label19
+            // 
+            label19.AutoSize = true;
+            label19.Location = new Point(39, 54);
+            label19.Name = "label19";
+            label19.Size = new Size(137, 15);
+            label19.TabIndex = 19;
+            label19.Text = "Verarbeitungsverzeichnis";
+            toolTip1.SetToolTip(label19, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // textBoxImageClassificationProcessingDirectory
+            // 
+            textBoxImageClassificationProcessingDirectory.Location = new Point(196, 51);
+            textBoxImageClassificationProcessingDirectory.Name = "textBoxImageClassificationProcessingDirectory";
+            textBoxImageClassificationProcessingDirectory.ReadOnly = true;
+            textBoxImageClassificationProcessingDirectory.Size = new Size(621, 23);
+            textBoxImageClassificationProcessingDirectory.TabIndex = 20;
+            // 
+            // buttonImageClassificationProcessingDirectory
+            // 
+            buttonImageClassificationProcessingDirectory.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
+            buttonImageClassificationProcessingDirectory.Location = new Point(823, 50);
+            buttonImageClassificationProcessingDirectory.Name = "buttonImageClassificationProcessingDirectory";
+            buttonImageClassificationProcessingDirectory.Size = new Size(123, 23);
+            buttonImageClassificationProcessingDirectory.TabIndex = 21;
+            buttonImageClassificationProcessingDirectory.Text = "Auswählen ...";
+            buttonImageClassificationProcessingDirectory.UseVisualStyleBackColor = true;
+            buttonImageClassificationProcessingDirectory.Click += buttonImageClassificationProcessingDirectory_Click;
+            // 
+            // label20
+            // 
+            label20.AutoSize = true;
+            label20.Location = new Point(39, 25);
+            label20.Name = "label20";
+            label20.Size = new Size(97, 15);
+            label20.TabIndex = 16;
+            label20.Text = "Dienstprogramm";
+            toolTip1.SetToolTip(label20, "In diesem Verzeichnis werden die zu verarbeitenden Dateien abgelegt");
+            // 
+            // textBoxImageClassificationServiceProgram
+            // 
+            textBoxImageClassificationServiceProgram.Location = new Point(196, 22);
+            textBoxImageClassificationServiceProgram.Name = "textBoxImageClassificationServiceProgram";
+            textBoxImageClassificationServiceProgram.ReadOnly = true;
+            textBoxImageClassificationServiceProgram.Size = new Size(621, 23);
+            textBoxImageClassificationServiceProgram.TabIndex = 17;
+            // 
+            // buttonImageClassificationServiceProgram
+            // 
+            buttonImageClassificationServiceProgram.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
+            buttonImageClassificationServiceProgram.Location = new Point(823, 21);
+            buttonImageClassificationServiceProgram.Name = "buttonImageClassificationServiceProgram";
+            buttonImageClassificationServiceProgram.Size = new Size(123, 23);
+            buttonImageClassificationServiceProgram.TabIndex = 18;
+            buttonImageClassificationServiceProgram.Text = "Auswählen ...";
+            buttonImageClassificationServiceProgram.UseVisualStyleBackColor = true;
+            buttonImageClassificationServiceProgram.Click += buttonImageClassificationServiceProgram_Click;
             // 
             // label_Spuren
             // 
@@ -444,7 +774,6 @@
             tableLayoutPanel1.Controls.Add(label_Warnung, 1, 4);
             tableLayoutPanel1.Controls.Add(button_DateienHinzufuegen, 3, 1);
             tableLayoutPanel1.Controls.Add(groupBox2, 1, 5);
-            tableLayoutPanel1.Controls.Add(groupBox1, 2, 5);
             tableLayoutPanel1.Dock = DockStyle.Fill;
             tableLayoutPanel1.Location = new Point(3, 3);
             tableLayoutPanel1.Margin = new Padding(0);
@@ -489,7 +818,7 @@
             // label1
             // 
             label1.AutoSize = true;
-            label1.Location = new Point(43, 43);
+            label1.Location = new Point(42, 71);
             label1.Name = "label1";
             label1.Size = new Size(107, 15);
             label1.TabIndex = 7;
@@ -499,7 +828,7 @@
             // label2
             // 
             label2.AutoSize = true;
-            label2.Location = new Point(43, 72);
+            label2.Location = new Point(42, 100);
             label2.Name = "label2";
             label2.Size = new Size(111, 15);
             label2.TabIndex = 10;
@@ -509,7 +838,7 @@
             // checkBoxAudioTranslationProcessVideos
             // 
             checkBoxAudioTranslationProcessVideos.AutoSize = true;
-            checkBoxAudioTranslationProcessVideos.Location = new Point(200, 98);
+            checkBoxAudioTranslationProcessVideos.Location = new Point(199, 126);
             checkBoxAudioTranslationProcessVideos.Name = "checkBoxAudioTranslationProcessVideos";
             checkBoxAudioTranslationProcessVideos.Size = new Size(123, 19);
             checkBoxAudioTranslationProcessVideos.TabIndex = 13;
@@ -517,6 +846,30 @@
             toolTip1.SetToolTip(checkBoxAudioTranslationProcessVideos, "Die Videoverarbeitung kann sehr lange dauern, wenn sich auf der Spur Kinofilme befinden");
             checkBoxAudioTranslationProcessVideos.UseVisualStyleBackColor = true;
             checkBoxAudioTranslationProcessVideos.CheckedChanged += checkBoxAudioTranslationProcessVideos_CheckedChanged;
+            // 
+            // checkBoxIPEDConfigEnableAudioTranslation
+            // 
+            checkBoxIPEDConfigEnableAudioTranslation.AutoSize = true;
+            checkBoxIPEDConfigEnableAudioTranslation.Location = new Point(199, 43);
+            checkBoxIPEDConfigEnableAudioTranslation.Name = "checkBoxIPEDConfigEnableAudioTranslation";
+            checkBoxIPEDConfigEnableAudioTranslation.Size = new Size(222, 19);
+            checkBoxIPEDConfigEnableAudioTranslation.TabIndex = 14;
+            checkBoxIPEDConfigEnableAudioTranslation.Text = "Audio Transkription und Übersetzung";
+            toolTip1.SetToolTip(checkBoxIPEDConfigEnableAudioTranslation, "Führt bei Audio- und Videodateien Transkriptionen und Übersetzung ins Deutsche mit Hilfe des Auswerteservers durch.");
+            checkBoxIPEDConfigEnableAudioTranslation.UseVisualStyleBackColor = true;
+            checkBoxIPEDConfigEnableAudioTranslation.CheckedChanged += checkBoxIPEDConfigEnabelAudioTranslation_CheckedChanged;
+            // 
+            // checkBoxIPEDConfigEnableImageClassification
+            // 
+            checkBoxIPEDConfigEnableImageClassification.AutoSize = true;
+            checkBoxIPEDConfigEnableImageClassification.Location = new Point(202, 43);
+            checkBoxIPEDConfigEnableImageClassification.Name = "checkBoxIPEDConfigEnableImageClassification";
+            checkBoxIPEDConfigEnableImageClassification.Size = new Size(125, 19);
+            checkBoxIPEDConfigEnableImageClassification.TabIndex = 20;
+            checkBoxIPEDConfigEnableImageClassification.Text = "Bilder klassifizieren";
+            toolTip1.SetToolTip(checkBoxIPEDConfigEnableImageClassification, "Führt eine Klassifizierung von Bilddateien mit Hilfe des Auswerteservers durch.");
+            checkBoxIPEDConfigEnableImageClassification.UseVisualStyleBackColor = true;
+            checkBoxIPEDConfigEnableImageClassification.CheckedChanged += checkBoxIPEDConfigEnableImageClassification_CheckedChanged;
             // 
             // tabControl1
             // 
@@ -699,6 +1052,8 @@
             // 
             // tabPageAudioTranslation
             // 
+            tabPageAudioTranslation.Controls.Add(groupBox3);
+            tabPageAudioTranslation.Controls.Add(checkBoxIPEDConfigEnableAudioTranslation);
             tabPageAudioTranslation.Controls.Add(checkBoxAudioTranslationProcessVideos);
             tabPageAudioTranslation.Controls.Add(label2);
             tabPageAudioTranslation.Controls.Add(textBoxAudioTranslationOutputDirectory);
@@ -716,7 +1071,7 @@
             // 
             // textBoxAudioTranslationOutputDirectory
             // 
-            textBoxAudioTranslationOutputDirectory.Location = new Point(200, 69);
+            textBoxAudioTranslationOutputDirectory.Location = new Point(199, 97);
             textBoxAudioTranslationOutputDirectory.Name = "textBoxAudioTranslationOutputDirectory";
             textBoxAudioTranslationOutputDirectory.ReadOnly = true;
             textBoxAudioTranslationOutputDirectory.Size = new Size(621, 23);
@@ -724,7 +1079,7 @@
             // 
             // textBoxAudioTranslationInputDirectory
             // 
-            textBoxAudioTranslationInputDirectory.Location = new Point(200, 40);
+            textBoxAudioTranslationInputDirectory.Location = new Point(199, 68);
             textBoxAudioTranslationInputDirectory.Name = "textBoxAudioTranslationInputDirectory";
             textBoxAudioTranslationInputDirectory.ReadOnly = true;
             textBoxAudioTranslationInputDirectory.Size = new Size(621, 23);
@@ -732,6 +1087,8 @@
             // 
             // tabPageImageClassification
             // 
+            tabPageImageClassification.Controls.Add(groupBox4);
+            tabPageImageClassification.Controls.Add(checkBoxIPEDConfigEnableImageClassification);
             tabPageImageClassification.Controls.Add(label4);
             tabPageImageClassification.Controls.Add(label3);
             tabPageImageClassification.Controls.Add(textBoxImageClassificationOutputDirectory);
@@ -748,7 +1105,7 @@
             // 
             // textBoxImageClassificationOutputDirectory
             // 
-            textBoxImageClassificationOutputDirectory.Location = new Point(200, 69);
+            textBoxImageClassificationOutputDirectory.Location = new Point(202, 97);
             textBoxImageClassificationOutputDirectory.Name = "textBoxImageClassificationOutputDirectory";
             textBoxImageClassificationOutputDirectory.ReadOnly = true;
             textBoxImageClassificationOutputDirectory.Size = new Size(621, 23);
@@ -756,7 +1113,7 @@
             // 
             // textBoxImageClassificationInputDirectory
             // 
-            textBoxImageClassificationInputDirectory.Location = new Point(200, 40);
+            textBoxImageClassificationInputDirectory.Location = new Point(202, 68);
             textBoxImageClassificationInputDirectory.Name = "textBoxImageClassificationInputDirectory";
             textBoxImageClassificationInputDirectory.ReadOnly = true;
             textBoxImageClassificationInputDirectory.Size = new Size(621, 23);
@@ -781,6 +1138,14 @@
             // 
             openFileDialogSettingsHashesDB.Filter = "IPED Hash Datenbankl|iped-hashes.db";
             // 
+            // openFileDialogAudioTranslationServiceProgram
+            // 
+            openFileDialogAudioTranslationServiceProgram.Filter = "Audiotranskription Dienstprogramml|background-media-translator.py";
+            // 
+            // openFileDialogImageClassificationServiceProgram
+            // 
+            openFileDialogImageClassificationServiceProgram.Filter = "Bildklassifizierung Dienstprogramml|image-classificator.py";
+            // 
             // MainForm
             // 
             AutoScaleDimensions = new SizeF(7F, 15F);
@@ -793,8 +1158,10 @@
             Text = "IPED GUI";
             groupBox2.ResumeLayout(false);
             groupBox2.PerformLayout();
-            groupBox1.ResumeLayout(false);
-            groupBox1.PerformLayout();
+            groupBox3.ResumeLayout(false);
+            groupBox3.PerformLayout();
+            groupBox4.ResumeLayout(false);
+            groupBox4.PerformLayout();
             tableLayoutPanel1.ResumeLayout(false);
             tableLayoutPanel1.PerformLayout();
             tabControl1.ResumeLayout(false);
@@ -834,8 +1201,6 @@
         private TabPage tabPageProtocol;
         private TabPage tabPageAudioTranslation;
         private TabPage tabPageImageClassification;
-        private CheckBox checkBoxIPEDConfigEnableAudioTranslation;
-        private CheckBox checkBoxIPEDConfigEnableImageClassification;
         private Label label2;
         private TextBox textBoxAudioTranslationOutputDirectory;
         private Label label1;
@@ -861,5 +1226,30 @@
         private OpenFileDialog openFileDialogSettingsIpedExePath;
         private FolderBrowserDialog folderBrowserDialog;
         private OpenFileDialog openFileDialogSettingsHashesDB;
+        private CheckBox checkBoxIPEDConfigEnableAudioTranslation;
+        private GroupBox groupBox3;
+        private Label label13;
+        private TextBox textBoxAudioTranslationFasterWhisperDirectory;
+        private Label label12;
+        private TextBox textBoxAudioTranslationProcessingDirectory;
+        private Label label11;
+        private TextBox textBoxAudioTranslationServiceProgram;
+        private ComboBox comboBoxAudioTranslationModelSize;
+        private Label label15;
+        private Label label14;
+        private TextBox textBoxAudioTranslationArgosTranslateDirectory;
+        private Button buttonAudioTranslationStartProcess;
+        private OpenFileDialog openFileDialogAudioTranslationServiceProgram;
+        private CheckBox checkBoxAudioTranslationUseGPU;
+        private CheckBox checkBoxIPEDConfigEnableImageClassification;
+        private Button buttonImageClassificationStartProcess;
+        private ComboBox comboBoxImageClassificationLanguage;
+        private Label label16;
+        private Label label18;
+        private TextBox textBoxImageClassificationMobileNetDirectory;
+        private Label label19;
+        private TextBox textBoxImageClassificationProcessingDirectory;
+        private TextBox textBoxImageClassificationServiceProgram;
+        private OpenFileDialog openFileDialogImageClassificationServiceProgram;
     }
 }